+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.4
+- Implement when and callback when finish setState.
+- Help to execute any other params when change state
+
 ## 1.0.3
 - Upgrade SDK.